Feature/enums (#444)
* * remove enums.ts

* * swap positions of STRING and NUMERIC_STRING
