+++ src/appliers/ifUndefined.ts
@@ -1,6 +1,5 @@
-import {CAUSE} from "../libs/enums";
 import {Key, Values} from "../libs/types";
-import {ValueSchemaError} from "../libs/ValueSchemaError";
+import {CAUSE, ValueSchemaError} from "../libs/ValueSchemaError";
 
 export interface Options<T>
 {
