+++ src/libs/constants.ts
@@ -1,18 +1,14 @@
-import {CHECKSUM_ALGORITHM, INTEGER} from "./enums";
 import {REGEXP_EMAIL} from "./regexp/email";
 import {REGEXP_IPV4} from "./regexp/ipv4";
 import {REGEXP_IPV6} from "./regexp/ipv6";
 import {REGEXP_HTTP, REGEXP_URI} from "./regexp/uri";
-import {CONVERT_CASE} from "../appliers/string/convertCase";
 
-export {CAUSE} from "./enums";
+import {INTEGER} from "../appliers/number/type";
+import {CONVERT_CASE} from "../appliers/string/convertCase";
+import {CHECKSUM_ALGORITHM} from "../appliers/numericString/checksum";
 
 export const NUMBER = {
 	INTEGER,
-};
-
-export const NUMERIC_STRING = {
-	CHECKSUM_ALGORITHM,
 };
 
 export const STRING = {
@@ -24,4 +20,8 @@ export const STRING = {
 		IPV6: REGEXP_IPV6,
 		URI: REGEXP_URI,
 	},
+};
+
+export const NUMERIC_STRING = {
+	CHECKSUM_ALGORITHM,
 };
